frontend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
python-docx
pillow
regex
pyyaml
